Machine_Learning/Collaborative_Filtering_Recommender_System/modular_code/requirements.txt
pandas == 1.3.5
scikit_learn == 0.24.1
scikit_surprise == 1.1.1
scipy == 1.6.2
